Expose Pipeline name & language to STT
Exposes the current assist_pipeline name and language to STT providers
via a modification to the stt.SpeechMetadata dataclass. Two new
empty-by-default fields are added to SpeechMetadata, which are unset.
Then before passing the audio stream to the STT provider, an override
SpeechMetadata instance is created which includes these fields, and this
override instance is then passed instead.

This is probably not the ideal way to do this, but is backwards
compatible with existing STT plugins.

The purpose of this is to expose these details to custom STT providers,
which may decide to alter behaviour based on the specific pipeline being
used (for instance, automatically determining a room/location based on
the pipeline chosen and altering the returned text to match).

Helps to implement the feature requested in:
https://community.home-assistant.io/t/bind-voice-assist-device-to-an-area/639998