feat: implement door_inverted binary sensor class
This class behaves identically to the door class, except its mapping of
on/off is inverted (on is closed, off is opened). Useful for a door,
presumably an interior door, which is normally-opened and where closed
in the "special" state. ("I'm studying with door closed, do X").

Implements https://community.home-assistant.io/t/flip-door-sensor-closed-state-color/868443